Feat: 저장된 비디오 삭제시 정말 삭제할거냐는 confirm 메시지 추가
Changes to be committed:
modified:   src/ts/components/SaveVideoSection.ts

+++ src/ts/components/SaveVideoSection.ts
@@ -13,6 +13,8 @@ const SaveVideoSection: Component<Props> = ({ children }) => {
 
   const onDeleteVideoHandler = ({ target }) => {
     if (!target.matches('.video-delete')) return;
+    if (!confirm('정말 삭제하시겠습니까?')) return;
+
     const videoId = target.parentNode.dataset.videoId;
 
     dispatch(videoDeleteAction(videoId));
